[Modify] #146 Room-Paging을 사용하기 위해 id값을 int형으로 변경
- int형만 지원

+++ data/src/main/java/com/woowahan/data/entity/table/OrderTableEntity.kt
@@ -12,7 +12,7 @@ data class OrderTableEntity(
     @ColumnInfo(name = OrderTableEntity.COLUMN_STATE)val deliveryState: Boolean = true
 ) {
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = OrderTableEntity.COLUMN_ID) var id: Long = 0
+    @ColumnInfo(name = OrderTableEntity.COLUMN_ID) var id: Int = 0
     companion object {
         const val TABLE_NAME = "order"
         const val COLUMN_ID = "id"
